external/webkit/LayoutTests/platform/android/webgl-conformance/dumpAsText/conformance/glsl/misc/shader-with-dfdx-no-ext.frag-expected.txt

CONSOLE MESSAGE: line 24: *** Error compiling shader '[object WebGLShader]':ERROR: 0:5: 'dFdx' : no matching overloaded function found 

fragment shader that uses dFdx without #extension should fail

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

test: fragment shader that uses dFdx without #extension should fail
default vertex shader
test fragment shaderPASS fragment shader that uses dFdx without #extension should fail
PASS successfullyParsed is true

TEST COMPLETE
